td-layout: Convert constants to u64
In order to avoid repeated casts to `u64` from types that are not directly
visible at that location (without looking at the source code), convert the
various layout constants to `u64` to start with.

If anything, it is a desirable property to require casts should these values be
used as `u32`.

Signed-off-by: Christophe de Dinechin <dinechin@redhat.com>